requirements.txt
# avro
auth0-python
channels
channels_redis
# confluent-kafka[avro]
Django>=3,<4
django-appconf
django-compressor
django-crispy-forms
django-libsass
django-prometheus
django-vouch-proxy-auth
# fastavro
pyyaml
gevent
gunicorn
psycopg2-binary
requests
uvicorn[standard]